+++ Slider/utils.js
@@ -3,6 +3,23 @@ import {is} from '@enact/core/keymap';
 import {clamp} from '@enact/core/util';
 import {calcProportion} from '@enact/ui/Slider/utils';
 
+const nop = () => {};
+
+const handleAcceleratedKeyDown = (ev, prop , {current: spotlightAccelerator}) => {
+	if (!spotlightAccelerator)
+		return true;
+
+	if (!ev.repeat) {
+		spotlightAccelerator.reset();
+	}
+
+	if (spotlightAccelerator.processKey(ev, nop)) {
+		return false;
+	}
+
+	return true;
+};
+
 const calcStep = (knobStep, step) => {
 	let s;
 
@@ -55,6 +72,7 @@ const handleIncrement = handle(
 	isIncrement,
 	preventDefault,
 	stop,
+	handleAcceleratedKeyDown,
 	isNotMax,
 	emitChange(1)
 );
@@ -64,6 +82,7 @@ const handleDecrement = handle(
 	isDecrement,
 	preventDefault,
 	stop,
+	handleAcceleratedKeyDown,
 	isNotMin,
 	emitChange(-1)
 );