+++ samples/sampler/stories/default/Slider.js
@@ -1,5 +1,5 @@
 import {action} from '@enact/storybook-utils/addons/actions';
-import {boolean, number, select} from '@enact/storybook-utils/addons/knobs';
+import {boolean, number, object, select} from '@enact/storybook-utils/addons/knobs';
 import {mergeComponentMetadata} from '@enact/storybook-utils';
 import {storiesOf} from '@storybook/react';
 
@@ -28,6 +28,7 @@ storiesOf('Sandstone', module)
 					activateOnSelect={boolean('activateOnSelect', SliderConfig) || false}
 					backgroundProgress={number('backgroundProgress', SliderConfig, {range: true, min: 0, max: 1, step: 0.01}, 0.5)}
 					disabled={disabled}
+					keyFrequency={object('keyFrequency', SliderConfig, [10])}
 					knobStep={number('knobStep', SliderConfig)}
 					max={number('max', SliderConfig, 10)}
 					min={number('min', SliderConfig, 0)}